Addressed review comments from Banks.
