+++ test/utils.js
@@ -1,0 +1,35 @@
+// Copyright 2019 OST.com Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+'use strict';
+
+function sleep(ms) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
+
+module.exports = {
+
+  sleep,
+
+  async waitBlockNumber(web3, blockNumber) {
+    let currentBlockNumber = await web3.eth.getBlockNumber();
+
+    while (currentBlockNumber < blockNumber) {
+      // eslint-disable-next-line no-await-in-loop
+      await sleep(200);
+      // eslint-disable-next-line no-await-in-loop
+      currentBlockNumber = await web3.eth.getBlockNumber();
+    }
+  },
+};
